[C10D] add _summarize_ranks util
Prints ranges of ranks succinctly.

e.g.

For a strided list of ranks, summarizes down to start:stop:step
```
0:4096:512
```

Omits step if it's 1
```
"0:8"
```

Note: endpoints are exclusive. This may not be intuitive to everyone,
but in the first above the last rank is 3584, and in the second it is
7.

Currently, does not support combinations of striding _and_ range.  (e.g.
can not generate a representation like "0:2, 4:6, ..., 12:14".  Is this
needed / useful? If so it could be added.

ghstack-source-id: 5a8f42160a708f0ea68717986a25127a2876674d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/160284