Add C++ function to accelerate DTensor.__new__
As the new comment explains, pybinding DispatchKeySet is expensive; one native function avoids the number of DispatchKeySets we incur pybind overhead for to 1.

Reviewers please advise on a better location for the new torch._C function; I just threw it in Module.cpp because I didn't know where to put it.

ghstack-source-id: 3633b3af68273f237acadab2618f637f7e496d2d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/161588